+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## Added
+- [Java] Add `Lineage.background()` and `.ruleBackground()`
 
 ## [14.7.0] - 2025-12-08
 ### Added