+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Removed
-- [JavaScript] Remove defunct legacy methods from `Query` ([#141](https://github.com/cucumber/query/pull/141))
+- [JavaScript] BREAKING CHANGE: Remove defunct legacy methods from `Query` ([#141](https://github.com/cucumber/query/pull/141))
 
 ## [14.7.0] - 2025-12-08
 ### Added